added config.load
