--- src/Assets/Bugsnag/Scripts/BugsnagSettingsObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a29d21abae70f4658920a0b2e27c189c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
