+++ src/BugsnagUnity/BugsnagSettingsObject.cs
@@ -29,9 +29,6 @@ namespace BugsnagUnity.Editor
             config.NotifyLogLevel = NotifyLogLevel;
             config.ReleaseStage = Debug.isDebugBuild ? "development" : "production";
             config.MaximumBreadcrumbs = MaximumBreadcrumbs;
-            config.ScriptingBackend = FindScriptingBackend();
-            config.DotnetScriptingRuntime = FindDotnetScriptingRuntime();
-            config.DotnetApiCompatibility = FindDotnetApiCompatibility();
             config.EnabledBreadcrumbTypes = GetEnabledBreadcrumbTypes();
             config.EnabledErrorTypes = EnabledErrorTypes;
             return config;
@@ -86,36 +83,6 @@ namespace BugsnagUnity.Editor
             public bool Error = true;
             public bool Manual = true;
         }
-        private static string FindScriptingBackend()
-        {
-#if ENABLE_MONO
-            return "Mono";
-#elif ENABLE_IL2CPP
-      return "IL2CPP";
-#else
-            return "Unknown";
-#endif
-        }
-
-        private static string FindDotnetScriptingRuntime()
-        {
-#if NET_4_6
-      return ".NET 4.6 equivalent";
-#else
-            return ".NET 3.5 equivalent";
-#endif
-        }
-
-        private static string FindDotnetApiCompatibility()
-        {
-#if NET_2_0_SUBSET
-      return ".NET 2.0 Subset";
-#else
-            return ".NET 2.0";
-#endif
-        }
-
-
 
     }
 }
